data-lander/requirements.txt
boto3
python-dateutil
dotenv